move hoc's into component folder
