+++ packages/pwa-kit-react-sdk/src/ssr/universal/components/with-error-handling/index.jsx
