+++ packages/pwa-kit-react-sdk/src/ssr/universal/components/with-loadable-resolver/index.jsx
